+++ server/pipeline_leaderboard.go
@@ -66,6 +66,21 @@ func (p *pipeline) leaderboardsList(logger *zap.Logger, session *session, envelo
 		params = append(params, incomingCursor.Id)
 	}
 
+	statements := make([]string, 0)
+	if len(incoming.GetFilterLeaderboardId()) != 0 {
+		for _, filterId := range incoming.GetFilterLeaderboardId() {
+			statement := "$" + strconv.Itoa(len(params))
+			statements = append(statements, statement)
+			params = append(params, filterId)
+		}
+
+		if len(incoming.Cursor) != 0 {
+			query += " AND "
+		}
+
+		query += "WHERE leaderboard.id IN (" + strings.Join(statements, ", ") + ")"
+	}
+
 	params = append(params, limit+1)
 	query += " LIMIT $" + strconv.Itoa(len(params))