+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 - Optionally send match data to a subset of match participants.
 - Fetch users by handle.
 - Add friend by handle. 
+- When listing leaderboards, optionally filter by ID.
 
 ### Changed
 - The build system now suffixes Windows binaries with `exe` extension.